requirements.txt
fastapi[all]
uvicorn
python-dotenv
httpx
transformer
torch


